Safari-609.1.11.1/LayoutTests/ietestcenter/Javascript/11.1.5_4-4-b-1-expected.txt

11.1.5_4-4-b-1 (TestCases/chapter11/11.1/11.1.5/11.1.5_4-4-b-1.js) Object literal - SyntaxError if a data property definition is followed by get accessor definition with the same name

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ES5Harness.preconditionPassed is true
FAIL ES5Harness.testPassed should be true (of type boolean). Was undefined (of type undefined).
PASS successfullyParsed is true

TEST COMPLETE
